Module js_builder::builder::expression::adapter

Adapters for transitioning between builders.

Structs

AssignAdapter
An adapter that holds the assignment expression type.

AssignRhsAdapter
An adapter that holds the type and left hand side of an assignment expression.

BinaryLhsAdapter
An adapter that holds the type of a binary operator.

CallAdapter
An adapter for function call expressions.

FnAccessAdapter
An adapter that receives the path to a function.

FnExprAdapter
An adapter that holds parts of a function expression.

FnExprDecl
The parts of a function expression.

GetterAdapter
An adapter that receives the name of an object field getter.

GetterBodyAdapter
An adapter that holds the name of an object field getter.

PropAssignAdapter
An adapter that receives the name of an object field (property).

PropInitAdapter
An adapter that holds the name of an object field (property).

SetterAdapter
An adapter that holds the argument, and receives the name of an object field setter.

SetterBodyAdapter
An adapter that holds the argument and the name of an object field setter.

UnaryAdapter
An adapter that holds the type of a unary operator.