Struct js_builder::builder::expression::ExpressionBuilder

#[must_use = "builders are sometimes lazy and must be used to produce a result"]
pub struct ExpressionBuilder<N: Build<Expression>>(_);

A builder for expressions.

Methods

impl<N: Build<Expression>> ExpressionBuilder<N>

fn this(self) -> N::Next

Build a this expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().this().compact_print(&mut result);

assert_eq!(
    "this",
    String::from_utf8_lossy(&result)
);

fn id<S: Into<Identifier>>(self, id: S) -> N::Next

Build an identifier expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().id("foo").compact_print(&mut result);

assert_eq!(
    "foo",
    String::from_utf8_lossy(&result)
);

fn ids<I: IntoIterator>(self, ids: I) -> N::Next where I::Item: Into<Identifier>

Build a path expression, consisting of only identifiers.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().ids(vec!["a", "b", "c"]).compact_print(&mut result);

assert_eq!(
    "a.b.c",
    String::from_utf8_lossy(&result)
);

fn null(self) -> N::Next

Build a null expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().null().compact_print(&mut result);

assert_eq!(
    "null",
    String::from_utf8_lossy(&result)
);

fn bool(self, b: bool) -> N::Next

Build a Boolean literal expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().bool(true).compact_print(&mut result);

assert_eq!(
    "true",
    String::from_utf8_lossy(&result)
);

fn number<Num: Into<f64>>(self, number: Num) -> N::Next

Build a number literal expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().number(3.14).compact_print(&mut result);

assert_eq!(
    "3.14",
    String::from_utf8_lossy(&result)
);

fn hex_int<I: Into<u64>>(self, int: I) -> N::Next

Build a hexadecimal integer expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().hex_int(0x1337beef_u64).compact_print(&mut result);

assert_eq!(
    "0x1337beef",
    String::from_utf8_lossy(&result)
);

fn string<S: Into<StringLiteral>>(self, string: S) -> N::Next

Build a string literal expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().string("Hello, JavaScript!").compact_print(&mut result);

assert_eq!(
    "'Hello, JavaScript!'",
    String::from_utf8_lossy(&result)
);

fn regex<S: Into<RegularExpression>>(self, regex: S) -> N::Next

Build a regular expression literal expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().regex(r"\w+").compact_print(&mut result);

assert_eq!(
    r"/\w+/",
    String::from_utf8_lossy(&result)
);

fn array(self) -> ArrayBuilder<Self>

Build an array literal expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().array()
    .elem().number(123)
    .elem().string("abc")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "[123, 'abc']",
    String::from_utf8_lossy(&result)
);

fn object(self) -> ObjectBuilder<Self>

Build an object literal expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().object()
    .field().id("a").number(123)
    .field().string("b").string("abc")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "{a: 123,'b': 'abc'}",
    String::from_utf8_lossy(&result)
);

fn paren(self) -> ExpressionBuilder<Self>

Put parenthesis around an expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().paren().string("wow").compact_print(&mut result);

assert_eq!(
    "('wow')",
    String::from_utf8_lossy(&result)
);

fn path(self) -> PathBaseBuilder<Self>

Build a path expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().path()
    .string("x")
    .id("y")
    .index().id("z")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "'x'.y[z]",
    String::from_utf8_lossy(&result)
);

fn new(self) -> PathBaseBuilder<FnAccessAdapter<Self>>

Build a constructor call expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().new().id("Point").build()
    .arg().id("x")
    .arg().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "new Point(x, y)",
    String::from_utf8_lossy(&result)
);

fn new_value(self) -> ValueExprBuilder<ValueExprBuilder<Self>>

Build a constructor call expression, without an argument list.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().new_value().path().id("Point").build().compact_print(&mut result);

assert_eq!(
    "new Point",
    String::from_utf8_lossy(&result)
);

fn function(self) -> FnArgnamesBuilder<FnExprAdapter<Self>>

Build a function expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().function()
    .arg("a")
    .arg("b")
    .build()
    .stmt().return_().add().id("a").id("b").build()
    .build()
    .compact_print(&mut result);

assert_eq!(
    "function(a, b) {return a + b;}",
    String::from_utf8_lossy(&result)
);

fn named_function<S: Into<Identifier>>(self, id: S) -> FnArgnamesBuilder<FnExprAdapter<Self>>

Build a named function expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().named_function("do_something")
    .arg("a")
    .arg("b")
    .build()
    .stmt().return_().add().id("a").id("b").build()
    .build()
    .compact_print(&mut result);

assert_eq!(
    "function do_something(a, b) {return a + b;}",
    String::from_utf8_lossy(&result)
);

fn assign(self) -> LhsExprBuilder<AssignAdapter<N>>

Build an assignment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().assign().id("a").id("b").compact_print(&mut result);

assert_eq!(
    "a = b",
    String::from_utf8_lossy(&result)
);

fn add_assign(self) -> LhsExprBuilder<AssignAdapter<N>>

Build an addition assignment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().add_assign().id("a").id("b").compact_print(&mut result);

assert_eq!(
    "a += b",
    String::from_utf8_lossy(&result)
);

fn sub_assign(self) -> LhsExprBuilder<AssignAdapter<N>>

Build a subtraction assignment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().sub_assign().id("a").id("b").compact_print(&mut result);

assert_eq!(
    "a -= b",
    String::from_utf8_lossy(&result)
);

fn mul_assign(self) -> LhsExprBuilder<AssignAdapter<N>>

Build a multiplication assignment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().mul_assign().id("a").id("b").compact_print(&mut result);

assert_eq!(
    "a *= b",
    String::from_utf8_lossy(&result)
);

fn div_assign(self) -> LhsExprBuilder<AssignAdapter<N>>

Build a division assignment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().div_assign().id("a").id("b").compact_print(&mut result);

assert_eq!(
    "a /= b",
    String::from_utf8_lossy(&result)
);

fn mod_assign(self) -> LhsExprBuilder<AssignAdapter<N>>

Build a modulo (remainder) assignment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().mod_assign().id("a").id("b").compact_print(&mut result);

assert_eq!(
    "a %= b",
    String::from_utf8_lossy(&result)
);

fn lsh_assign(self) -> LhsExprBuilder<AssignAdapter<N>>

Build a left shift assignment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().lsh_assign().id("a").id("b").compact_print(&mut result);

assert_eq!(
    "a <<= b",
    String::from_utf8_lossy(&result)
);

fn rsh_assign(self) -> LhsExprBuilder<AssignAdapter<N>>

Build a right shift assignment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().rsh_assign().id("a").id("b").compact_print(&mut result);

assert_eq!(
    "a >>= b",
    String::from_utf8_lossy(&result)
);

fn ursh_assign(self) -> LhsExprBuilder<AssignAdapter<N>>

Build an unsigned right shift assignment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().ursh_assign().id("a").id("b").compact_print(&mut result);

assert_eq!(
    "a >>>= b",
    String::from_utf8_lossy(&result)
);

fn and_assign(self) -> LhsExprBuilder<AssignAdapter<N>>

Build a bitwise and assignment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().and_assign().id("a").id("b").compact_print(&mut result);

assert_eq!(
    "a &= b",
    String::from_utf8_lossy(&result)
);

fn or_assign(self) -> LhsExprBuilder<AssignAdapter<N>>

Build a bitwise or assignment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().or_assign().id("a").id("b").compact_print(&mut result);

assert_eq!(
    "a |= b",
    String::from_utf8_lossy(&result)
);

fn xor_assign(self) -> LhsExprBuilder<AssignAdapter<N>>

Build a bitwise xor assignment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().xor_assign().id("a").id("b").compact_print(&mut result);

assert_eq!(
    "a ^= b",
    String::from_utf8_lossy(&result)
);

fn add(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build an addition expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().add().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a + b + c",
    String::from_utf8_lossy(&result)
);

fn sub(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a subtraction expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().sub().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a - b - c",
    String::from_utf8_lossy(&result)
);

fn mul(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a multiplication expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().mul().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a * b * c",
    String::from_utf8_lossy(&result)
);

fn div(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a division expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().div().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a / b / c",
    String::from_utf8_lossy(&result)
);

fn mod_(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a modulo (reminder) expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().mod_().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a % b % c",
    String::from_utf8_lossy(&result)
);

fn lsh(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a left shift expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().lsh().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a << b << c",
    String::from_utf8_lossy(&result)
);

fn rsh(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a right shift expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().rsh().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a >> b >> c",
    String::from_utf8_lossy(&result)
);

fn ursh(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build an unsigned right shift expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().ursh().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a >>> b >>> c",
    String::from_utf8_lossy(&result)
);

fn lt(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a less than expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().lt().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a < b < c",
    String::from_utf8_lossy(&result)
);

fn gt(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a greater than expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().gt().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a > b > c",
    String::from_utf8_lossy(&result)
);

fn lt_eq(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a less than or equal expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().lt_eq().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a <= b <= c",
    String::from_utf8_lossy(&result)
);

fn gt_eq(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a greater than or equal expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().gt_eq().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a >= b >= c",
    String::from_utf8_lossy(&result)
);

fn eq(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build an equality expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().eq().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a == b == c",
    String::from_utf8_lossy(&result)
);

fn not_eq(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build an inequality expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().not_eq().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a != b != c",
    String::from_utf8_lossy(&result)
);

fn identical(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build an identity expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().identical().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a === b === c",
    String::from_utf8_lossy(&result)
);

fn not_identical(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a nonidentity expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().not_identical().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a !== b !== c",
    String::from_utf8_lossy(&result)
);

fn and(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a logical and expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().and().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a && b && c",
    String::from_utf8_lossy(&result)
);

fn or(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a logical or expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().or().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a || b || c",
    String::from_utf8_lossy(&result)
);

fn bit_and(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a bitwise and expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().bit_and().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a & b & c",
    String::from_utf8_lossy(&result)
);

fn bit_or(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a bitwise or expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().bit_or().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a | b | c",
    String::from_utf8_lossy(&result)
);

fn bit_xor(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build a bitwise xor expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().bit_xor().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a ^ b ^ c",
    String::from_utf8_lossy(&result)
);

fn instanceof(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build an instanceof expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().instanceof().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a instanceof b instanceof c",
    String::from_utf8_lossy(&result)
);

fn in_(self) -> ExpressionBuilder<BinaryLhsAdapter<N>>

Build an in expression. At least two sub-expressions are required, and more can be added. See BinaryExprBuilder for more info.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().in_().id("a").id("b").expr().id("c").build().compact_print(&mut result);

assert_eq!(
    "a in b in c",
    String::from_utf8_lossy(&result)
);

fn pre_incr(self) -> LhsExprBuilder<UnaryAdapter<N>>

Build a prefix increment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().pre_incr().id("a").compact_print(&mut result);

assert_eq!(
    "++a",
    String::from_utf8_lossy(&result)
);

fn post_incr(self) -> LhsExprBuilder<UnaryAdapter<N>>

Build a postfix increment expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().post_incr().id("a").compact_print(&mut result);

assert_eq!(
    "a++",
    String::from_utf8_lossy(&result)
);

fn pre_decr(self) -> LhsExprBuilder<UnaryAdapter<N>>

Build a prefix decrement expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().pre_decr().id("a").compact_print(&mut result);

assert_eq!(
    "--a",
    String::from_utf8_lossy(&result)
);

fn post_decr(self) -> LhsExprBuilder<UnaryAdapter<N>>

Build a postfix decrement expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().post_decr().id("a").compact_print(&mut result);

assert_eq!(
    "a--",
    String::from_utf8_lossy(&result)
);

fn delete(self) -> LhsExprBuilder<UnaryAdapter<N>>

Build a delete expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().delete().id("a").compact_print(&mut result);

assert_eq!(
    "delete a",
    String::from_utf8_lossy(&result)
);

fn void(self) -> LhsExprBuilder<UnaryAdapter<N>>

Build a void expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().void().id("a").compact_print(&mut result);

assert_eq!(
    "void a",
    String::from_utf8_lossy(&result)
);

fn typeof_(self) -> LhsExprBuilder<UnaryAdapter<N>>

Build a typeof expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().typeof_().id("a").compact_print(&mut result);

assert_eq!(
    "typeof a",
    String::from_utf8_lossy(&result)
);

fn plus(self) -> LhsExprBuilder<UnaryAdapter<N>>

Build a unary plus expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().plus().id("a").compact_print(&mut result);

assert_eq!(
    "+a",
    String::from_utf8_lossy(&result)
);

fn neg(self) -> LhsExprBuilder<UnaryAdapter<N>>

Build a negation expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().neg().id("a").compact_print(&mut result);

assert_eq!(
    "-a",
    String::from_utf8_lossy(&result)
);

fn not(self) -> LhsExprBuilder<UnaryAdapter<N>>

Build a logical not expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().not().id("a").compact_print(&mut result);

assert_eq!(
    "!a",
    String::from_utf8_lossy(&result)
);

fn bit_not(self) -> LhsExprBuilder<UnaryAdapter<N>>

Build a bitwise not expression.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().bit_not().id("a").compact_print(&mut result);

assert_eq!(
    "~a",
    String::from_utf8_lossy(&result)
);

fn build_expr(self, expression: Expression) -> N::Next

Use a pre-built expression, instead.

impl<N: Build<Expression> + Build<Expressions>> ExpressionBuilder<N>

fn comma(self) -> ExpressionsBuilder<N>

Build a comma separated expression chain.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().comma()
    .expr().id("a")
    .expr().id("b")
    .expr().id("c")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "a, b, c",
    String::from_utf8_lossy(&result)
);