Struct js_builder::builder::expression::PathBaseBuilder

#[must_use = "builders are sometimes lazy and must be used to produce a result"]
pub struct PathBaseBuilder<N: Build<PathExpression>>(_);

A builder for the first part of a path.

Methods

impl<N: Build<PathExpression>> PathBaseBuilder<N>

fn this(self) -> PathBuilder<PathExpression, N>

Build a path that begins with this.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().this()
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "this.x[y]",
    String::from_utf8_lossy(&result)
);

fn id<S: Into<Identifier>>(self, id: S) -> PathBuilder<PathExpression, N>

Build a path that begins with an identifier.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().id("foo")
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "foo.x[y]",
    String::from_utf8_lossy(&result)
);

fn null(self) -> PathBuilder<PathExpression, N>

Build a path that begins with null.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().null()
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "null.x[y]",
    String::from_utf8_lossy(&result)
);

fn bool(self, b: bool) -> PathBuilder<PathExpression, N>

Build a path that begins with a Boolean value.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().bool(false)
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "false.x[y]",
    String::from_utf8_lossy(&result)
);

fn number<Num: Into<f64>>(self, number: Num) -> PathBuilder<PathExpression, N>

Build a path that begins with a number. You may want to wrap the number in parenthesis first, to avoid ambiguities. See paren, below.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().number(3.14)
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "3.14.x[y]",
    String::from_utf8_lossy(&result)
);

fn hex_int<I: Into<u64>>(self, int: I) -> PathBuilder<PathExpression, N>

Build a path that begins with a hexadecimal integer.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().hex_int(0x1337beef_u64)
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "0x1337beef.x[y]",
    String::from_utf8_lossy(&result)
);

fn string<S: Into<StringLiteral>>(self, string: S) -> PathBuilder<PathExpression, N>

Build a path that begins with a string.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().string("hello")
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "'hello'.x[y]",
    String::from_utf8_lossy(&result)
);

fn regex<S: Into<RegularExpression>>(self, regex: S) -> PathBuilder<PathExpression, N>

Build a path that begins with a regular expression.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().regex(r"\w+")
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    r"/\w+/.x[y]",
    String::from_utf8_lossy(&result)
);

fn array(self) -> ArrayBuilder<Self>

Build a path that begins with an array.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().array()
    .elem().id("a")
    .elem().id("b")
    .elem().id("c")
    .build()
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    r"[a, b, c].x[y]",
    String::from_utf8_lossy(&result)
);

fn object(self) -> ObjectBuilder<Self>

Build a path that begins with an object.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().object()
    .field().id("a").number(1)
    .field().id("b").number(2)
    .field().id("c").number(3)
    .build()
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    r"{a: 1,b: 2,c: 3}.x[y]",
    String::from_utf8_lossy(&result)
);

fn paren(self) -> ExpressionBuilder<Self>

Build a path that begins with an expression in parenthesis.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().paren().number(4)
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "(4).x[y]",
    String::from_utf8_lossy(&result)
);

fn function(self) -> FnArgnamesBuilder<FnExprAdapter<Self>>

Build a path that begins with a function expression.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().function()
    .arg("a")
    .arg("b")
    .build()
    .stmt().return_().add().id("a").id("b").build()
    .build()
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "(function(a, b) {return a + b;}).x[y]",
    String::from_utf8_lossy(&result)
);

fn named_function<S: Into<Identifier>>(self, id: S) -> FnArgnamesBuilder<FnExprAdapter<Self>>

Build a path that begins with a function expression.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().named_function("do_something")
    .arg("a")
    .arg("b")
    .build()
    .stmt().return_().add().id("a").id("b").build()
    .build()
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "(function do_something(a, b) {return a + b;}).x[y]",
    String::from_utf8_lossy(&result)
);

fn new(self) -> PathBaseBuilder<FnAccessAdapter<Self>>

Build a path that begins with a constructor call.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().new().id("Foo").build()
    .arg().id("a")
    .arg().id("b")
    .build()
    .id("x")
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "new Foo(a, b).x[y]",
    String::from_utf8_lossy(&result)
);

fn ids<I: IntoIterator>(self, ids: I) -> PathBuilder<PathExpression, N> where I::Item: Into<Identifier>

Build a path that begins with a sequence of identifiers.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().ids(vec!["foo", "x"])
    .index().id("y")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "foo.x[y]",
    String::from_utf8_lossy(&result)
);