Struct js_builder::builder::expression::PathBuilder

#[must_use = "builders are sometimes lazy and must be used to produce a result"]
pub struct PathBuilder<E: PathType, N: Build<E>>(_, _);

A builder for path expressions.

Methods

impl<E: PathType, N: Build<E>> PathBuilder<E, N>

fn id<S: Into<Identifier>>(self, id: S) -> PathBuilder<E, N>

Access a member, using the dot notation.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().id("a")
    .id("b")
    .id("c")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "a.b.c",
    String::from_utf8_lossy(&result)
);

fn add_id<S: Into<Identifier>>(&mut self, id: S) -> &mut PathBuilder<E, N>

Same as id, but without moving the builder.

fn index(self) -> ExpressionBuilder<Self>

Access a member, using the bracket notation.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().id("a")
    .index().id("b")
    .index().id("c")
    .build()
    .compact_print(&mut result);

assert_eq!(
    "a[b][c]",
    String::from_utf8_lossy(&result)
);

fn add_index(&mut self) -> ExpressionBuilder<&mut Self>

Same as index, but without moving the builder.

fn with_index(&mut self, expr: Expressions) -> &mut PathBuilder<E, N>

Index the path, using a pre-built expression chain.

fn call(self) -> FnCallBuilder<E, CallAdapter<N>> where N: Build<CallExpression>

Call the current path as a function, and continue building it.

use js_builder::ast::PathExpression;
use js_builder::Print;

let mut result = Vec::new();
PathExpression::build().id("a")
    .id("b")
    .id("c")
    .call()
    .arg().id("x")
    .arg().id("y")
    .build()
    .id("toString")
    .call().build()
    .build()
    .compact_print(&mut result);

assert_eq!(
    "a.b.c(x, y).toString()",
    String::from_utf8_lossy(&result)
);

fn build(self) -> N::Next

Finish building the path.