Struct js_builder::builder::function::FnArgnamesBuilder

#[must_use = "builders are sometimes lazy and must be used to produce a result"]
pub struct FnArgnamesBuilder<N: Build<Vec<Identifier>>>(_, _);

A builder for function argument names.

Methods

impl<N: Build<Vec<Identifier>>> FnArgnamesBuilder<N>

fn arg<S: Into<Identifier>>(self, id: S) -> FnArgnamesBuilder<N>

Add an argument to the argument list.

use js_builder::ast::Expression;
use js_builder::Print;

let mut result = Vec::new();
Expression::build().function()
    .arg("a")
    .arg("b")
    .arg("c")
    .build() //finish the argument list and start building the function body
    .stmt().expr().assign().id("a").add().id("b").id("c").build()
    .build()
    .compact_print(&mut result);

assert_eq!(
    "function(a, b, c) {a = b + c;}",
    String::from_utf8_lossy(&result)
);

fn add_arg<S: Into<Identifier>>(&mut self, id: S) -> &mut FnArgnamesBuilder<N>

Same as arg, but without moving the builder.

fn build(self) -> N::Next

Finish building the function argument list.