Struct js_builder::builder::function::FnCallBuilder

#[must_use = "builders are sometimes lazy and must be used to produce a result"]
pub struct FnCallBuilder<E: PathType, N: Build<FnCall<E>>> {
    // some fields omitted
}

A builder for function calls.

Methods

impl<E: PathType, N: Build<FnCall<E>>> FnCallBuilder<E, N>

fn arg(self) -> ExpressionBuilder<Self>

Add an argument to the function call.

fn add_arg(&mut self) -> ExpressionBuilder<&mut Self>

Same as arg, but without moving the builder.

fn with_arg(&mut self, arg: Expression) -> &mut FnCallBuilder<E, N>

Add a pre-built expression as an argument.

fn build(self) -> N::Next

Finish build the function call.

fn from(path: E, next: N) -> FnCallBuilder<E, N>

Create a new function call builder from a path and a builder.