Module js_builder::builder::statement::adapter

Adapters for transitioning between builders.

Structs

DoAdapter
Adapter for receiving the do ... statement in a do ... while(...) loop.

DoWhileAdapter
Adapter for receiving the while(...) condition in a do ... while(...) loop.

ForCondAdapter
Adapter for receiving the condition expression of a for loop.

ForInLhsAdapter
An adapter for the left hand side of a for(... in ...) statement.

ForInRhsAdapter
An adapter for the right hand side of a for(... in ...) statement.

ForInStmtAdapter
An adapter for the loop body of for(... in ...) loops.

ForIncrAdapter
Adapter for receiving the incrementation expression of a for loop.

ForInitAdapter
Adapter for receiving the initiation statement of a for loop.

ForStmtAdapter
Adapter for receiving the body statement of a for loop.

IfAdapter
An adapter for the condition expression in an if statement.

InitAdapter
An adapter that holds the name of a variable.

ReturnAdapter
An expression adapter for return statements.

ThenAdapter
An adapter for holding the if condition while the first branch is built.

WhileAdapter
Adapter for receiving the while(...) condition in a while(...) ... loop.

WhileDoAdapter
Adapter for receiving the loop body in a while(...) ... loop.