Struct js_builder::builder::statement::adapter::ForInRhsAdapter

pub struct ForInRhsAdapter<N: Build<Statement>>(_, _);

An adapter for the right hand side of a for(... in ...) statement.

Methods

impl<N: Build<Statement>> ForInRhsAdapter<N>

fn from<S: Into<Identifier>>(id: S, next: N) -> ForInRhsAdapter<N>

Create a for(var ... in ...) RHS expression adapter from a variable name and a builder.